fix to seal selection when seals become unavailable
The seal picker had an issue where if a seal available through a talent or rune became no longer available, it defaulted to NoSeal leading to issues in the sim backend.

This is now accounted for in the backend via tweaks to the primary seal APLAction and how the APLValue for remaining seal time is set on reset to be very very large to prevent checks on the value looping.

Also included a warning in the sim UI for when no seal is selected, warning the user that their previously selected seal is no longer available, that no seal will be cast in sims of the current config, and prompting them to select a valid seal