Fixes to judgement spells
Before I had incorrectly assumed that all dps judgements in classic operated via an intermediary dummy spell that rolled spellhit; this is only true for Judgement of Command.

This commit refactors judges to ensure JoR/JoM are directly called as in the game.

JoC is still called via the intermediary judge such that the dummy can roll on spell hit, and the actual JoC command can roll special melee crit/hit.

The judgement dummy spell is displayed in the Damage results for SoC so as to give the user an idea of the miss rate for their judgement spell; hiding this spell via the SpellFlagsNoMetrics spell might mislead the user into thinking their Judgement of Command spell cannot miss, when technically the intermediary dummy spell can.